Major restructure of nav and header.
Header is an fixed position element that covers the entire screen with an opacity of 95%. Tagline displays based on conditional logic: only visible on home page. (There is some latency issue with this. The tagline sticks around for a half second or so when you go to one of the non-home pages.)