Implement dynamically updating line chart.
Abstract D3 out of the directive. And abstract data generation out of the directive into the factory.

Use OOP for the directive and factory.

Change the factory to a service. It seems that factories are for functional programming folks, and services are for OOPs.

This work is the result of my session with Jacob Morris of Pivotal Labs. Thanks, Jacob!